backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.65.0
pydantic==2.10.0
pydantic-settings==2.8.0
pypdf==5.3.0
pytesseract==0.3.13
pillow==11.0.0
docx2txt==0.8 